requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Core Airflow providers
apache-airflow-providers-postgres==5.10.0

# Data processing
pandas==2.0.3
requests==2.31.0

# Version control
dvc==3.30.0
gitpython==3.1.40

# Database
psycopg2-binary==2.9.9

# Utility
python-dotenv==1.0.0